+++ actiondb-parser/actiondb/src/matcher/pattern/file/mod.rs
@@ -1,5 +1,3 @@
-pub use self::file::PatternFile;
-
 mod deser;
 mod file;
 mod iter;
